+++ backend/open_webui/models/auths.py
@@ -87,8 +87,8 @@ class SignupForm(BaseModel):
     name: str
     email: str
     password: str
-    charity_id: Optional[int] = None
     profile_image_url: Optional[str] = "/user.png"
+    charity_id: Optional[int] = None
 
 
 class AddUserForm(SignupForm):
