+++ backend/open_webui/models/users.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from sqlalchemy.orm import relationship, selectinload
 
-from open_webui.models.profiles import UserProfile
+from open_webui.models.profiles import UserProfile, UserProfileModel
 from open_webui.internal.db import Base, JSONField, get_db
 
 
@@ -48,18 +48,6 @@ class UserSettings(BaseModel):
     model_config = ConfigDict(extra="allow")
     pass
 
-
-class CharityModel(BaseModel):
-    id: int
-    name: str
-    charity_id: Optional[int] = None
-
-    model_config = ConfigDict(from_attributes=True)
-
-class UserProfileModel(BaseModel):
-    charity: Optional[CharityModel] = None
-
-    model_config = ConfigDict(from_attributes=True)
 
 class UserModel(BaseModel):
     id: str
@@ -421,26 +409,5 @@ class UsersTable:
                 return UserModel.model_validate(user)
             else:
                 return None
-
-    def set_user_charity(self, user_id: str, charity_id: Optional[int]) -> bool:
-        """
-        Set the user's charity. Creates a UserProfile for the user if it does not exist.
-        """
-        try:
-            with get_db() as db:
-                user = db.query(User).filter_by(id=user_id).first()
-                if not user:
-                    return False
-
-                # Ensure UserProfile exists
-                if not user.profile:
-                    user.profile = UserProfile(user_id=user.id)
-
-                user.profile.charity_id = charity_id
-                db.commit()
-                return True
-        except Exception as e:
-            print("Error in set_user_charity:", e)
-            return False
 
 Users = UsersTable()
