+++ src/lib/apis/auths/index.ts
@@ -291,7 +291,7 @@ export const userSignUp = async (
 	email: string,
 	password: string,
 	profile_image_url: string,
-	charity?: number, // optional
+	charity?: number // optional
 ) => {
 	let error = null;
 
@@ -304,7 +304,7 @@ export const userSignUp = async (
 		body: JSON.stringify({
 			name: name,
 			email: email,
-      charity_id: charity,
+			charity_id: charity,
 			password: password,
 			profile_image_url: profile_image_url
 		})
@@ -361,7 +361,7 @@ export const addUser = async (
 	password: string,
 	role: string = 'pending',
 	profile_image_url: null | string = null,
-  charity_id: null | number = null
+	charity_id: null | number = null
 ) => {
 	let error = null;
 
@@ -377,7 +377,7 @@ export const addUser = async (
 			password: password,
 			role: role,
 			...(profile_image_url && { profile_image_url: profile_image_url }),
-      charity_id: charity_id,
+			charity_id: charity_id
 		})
 	})
 		.then(async (res) => {
