+++ src/lib/components/admin/Users/UserList/AddUserModal.svelte
@@ -6,7 +6,7 @@
 
 	import { WEBUI_BASE_URL } from '$lib/constants';
 
-  import CharityAutocomplete from '$lib/components/CharityAutocomplete.svelte';
+	import CharityAutocomplete from '$lib/components/CharityAutocomplete.svelte';
 	import Spinner from '$lib/components/common/Spinner.svelte';
 	import Modal from '$lib/components/common/Modal.svelte';
 	import { generateInitialsImage } from '$lib/utils';
@@ -20,14 +20,14 @@
 	let loading = false;
 	let tab = '';
 	let inputFiles;
-  let selectedCharity = null;
+	let selectedCharity = null;
 
 	let _user = {
 		name: '',
 		email: '',
 		password: '',
 		role: 'user',
-    profile: null
+		profile: null
 	};
 
 	$: if (show) {
@@ -36,7 +36,7 @@
 			email: '',
 			password: '',
 			role: 'user',
-			profile: null,
+			profile: null
 		};
 	}
 
@@ -56,7 +56,7 @@
 				_user.password,
 				_user.role,
 				generateInitialsImage(_user.name),
-        selectedCharity?.id,
+				selectedCharity?.id
 			).catch((error) => {
 				toast.error(`${error}`);
 			});
@@ -198,11 +198,9 @@
 								<div class=" mb-1 text-xs text-gray-500">{$i18n.t('Charity')}</div>
 
 								<div class="flex-1">
-                  <CharityAutocomplete bind:value={selectedCharity} />
-
+									<CharityAutocomplete bind:value={selectedCharity} />
 								</div>
 							</div>
-
 
 							<div class="flex flex-col w-full mt-1">
 								<div class=" mb-1 text-xs text-gray-500">{$i18n.t('Name')}</div>
