+++ src/lib/components/admin/Users/UserList/EditUserModal.svelte
@@ -6,7 +6,7 @@
 
 	import { updateUserById } from '$lib/apis/users';
 
-  import CharityAutocomplete from '$lib/components/CharityAutocomplete.svelte';
+	import CharityAutocomplete from '$lib/components/CharityAutocomplete.svelte';
 	import Modal from '$lib/components/common/Modal.svelte';
 	import localizedFormat from 'dayjs/plugin/localizedFormat';
 	import XMark from '$lib/components/icons/XMark.svelte';
@@ -25,16 +25,16 @@
 		name: '',
 		email: '',
 		password: '',
-    profile: null
+		profile: null
 	};
 
-  let selectedCharity = null;
+	let selectedCharity = null;
 
 	const submitHandler = async () => {
-    // User profile must exist before the assignment of selectedCharity
-    if (!_user.profile) _user.profile = {};
+		// User profile must exist before the assignment of selectedCharity
+		if (!_user.profile) _user.profile = {};
 
-   _user.profile.charity = selectedCharity; 
+		_user.profile.charity = selectedCharity;
 
 		const res = await updateUserById(localStorage.token, selectedUser.id, _user).catch((error) => {
 			toast.error(`${error}`);
@@ -46,16 +46,13 @@
 		}
 	};
 
-  onMount(() => {
+	onMount(() => {
 		if (selectedUser) {
 			_user = selectedUser;
 			_user.password = '';
-      selectedCharity = _user.profile?.charity || null;
-    }
-  });
-
-
-
+			selectedCharity = _user.profile?.charity || null;
+		}
+	});
 </script>
 
 <Modal size="sm" bind:show>
@@ -122,9 +119,9 @@
 								<div class=" mb-1 text-xs text-gray-500">{$i18n.t('Charity')}</div>
 
 								<div class="flex-1">
-                  <CharityAutocomplete bind:value={selectedCharity} />
-                </div>
-              </div>
+									<CharityAutocomplete bind:value={selectedCharity} />
+								</div>
+							</div>
 
 							<div class="flex flex-col w-full">
 								<div class=" mb-1 text-xs text-gray-500">{$i18n.t('Email')}</div>
